Minimal file required to use @turf/area module
Use this script to calculate area of a polygon on the earth's surface.

The file is composed from Turf.js (http://turfjs.org/) - Advanced geospatial analysis for browsers and Node.js

The complete Turf.js distribution is about 500kb. 

Only the area and polygon modules have been included in this file.